Added a Death singleton.  Whenever an actor dies it logs it and the day they died.  Usefull for debugging can check if something is trying to interact with a dead actor.
EconomicActorManager is now passing it self to the actors.  Moving
away from the ActorAction enum thing.  The manager has a FindPerson
and FindCompany method that the actors can use to find the person
they are interacting wth.  Not all the way there yet.  Still
need to remove all references to the fConnection thing so it can
be removed